backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0
gunicorn==21.2.0
requests==2.31.0